Add attributes to enum values (#86)
* Add attributes to enum values

As of this change, the grammar allows attributes on enum values, either
inline or indented in the body, the same as fields. Additionally, this
change updates the formatter and grammar documentation to account for
this change.

* Check attributes of enum values

This ensures that enum values are checked during attribute verification.
Since currently there are no non-backend attributes for enum values,
having any (non-backend) attribute at all will produce an error.