Emit compilation errors for bad C++ namespaces
Prior to this CL, an incorrectly-formatted namespace will fail an assert
and print a python exception to the user. While this likely sufficient
for a user to diagnose the issue, it is inconsistent with other error
reporting in Emboss.

With this change, the namespace checking is implemented in the same way
as other validations in the Emboss compiler, providing a specific file,
line, and column range of the issue and printing the line containing the
issue to the user.