Add forward declarations for text helpers
This moves text related helpers out to the `emboss::support` namespace
in `emboss_text_util.h`. We add forward declarations for the helpers and
convert a few options instances to take refs instead. This is in
preparation for adding support for disabling the generation of
text-related helpers such as `EnumTraits`.