Error example enumerator for Emboss.
This is a program that generates examples of erroneous parses, along
with some information about the potential errors, with the intention of
allowing developers to quickly find out what states do *not* have
associated error messages.

Ideally, this program would be used to check that there *are no* states
that are missing error messages, but actually adding all of those
messages is not feasible right now: at this change,
enumerate_parse_errors finds 5808 states that are missing messages.