Add build.json and generator script
This commit introduces a new script, `manage_build_json.py`, which
queries Bazel to determine the source files for the Emboss compiler
and C++ runtime.

The script generates a `build.json` file in the project root, which
will serve as the source of truth for generating build helper files
for other build systems.

The script also adds support for verifying that `build.json` is in
sync with the bazel build targets.