Add build files and generator script
Adds a `generate_build_files.py` script that is used to generate build
files for use in bazel, cmake, and GN. Additionally a clean json file is
added for more generic usage. These files are intended to be included in
downstream build systems to explicitly list out the files required by
embossc and the C++ runtime.

The generated build files are included as well.